SCALE-SEAM ⑦: the /auth group out of client.ts, and the group the roadmap said would be hard was not
Route-group `/auth` moves to `api/auth.ts` as a mixin — 20 methods / 96 lines, `client.ts`
3,968 → 3,872. Located by the route each method calls, bodies by brace matching, doc comments
claimed by scanning backwards. They sat in four separate regions of the file.

**The predicted difficulty was already discharged.** ⑥ flagged `/auth` as the one group that "owns
token state rather than just calling routes". It does mutate token state — `changePassword` and
`logoutAll` both adopt the fresh token the server returns — but that state has lived on `HttpCore`
behind a public `setToken` since the T2 transport extraction. A mixin is a BASE of `ApiClient`: it
cannot see `ApiClient`'s privates, but its own base's public members are fine. The blocker that
actually stopped the SSE methods travelling in ③ was `liveStream` being private *on ApiClient*, and
there is no analogue here. `token` stays private on `HttpCore` and is untouched.

**Four methods deliberately did not move.** `auditLog`, `errorLog`, `clearErrorLog` and
`reportClientError` sit inside the `// --- admin: user management ---` run and read as part of this
group, but they route to `/audit`, `/admin/errors` and `/client-errors`. Grouping by the section
comment instead of the route would have dragged three unrelated domains across the seam — which is
the whole reason the recipe is "locate by route". The now-orphaned `// --- auth ---` banner is
deleted rather than left labelling `integrations()`.

**The surface ratchet had drifted back into slack, and only measuring found it.** Floor was 698;
this reader counts 699. Raised to 699. The evidence that the move dropped nothing is that the count
was 699 **before** the extraction as well as after — measured on the unmodified tree with the gate's
own reader, not a probe of mine. So the slack was not created here; it accumulated from merges since
⑥, exactly as that test's comment predicted every merge would.

Mutation-checked both directions, each verified to have actually applied before the result was read
(the first attempt's regex silently matched nothing under CRLF and reported a meaningless green):
deleting only `login` → 698, fails by number and names `login`; leaving the mixin uncomposed → 679,
fails by 20. At the old 698 floor the single-method loss would have passed silently, which is why
the auth names are now in the spot-check list too — losing all 20 was always loud, losing one was not.

tsc clean, eslint clean, web suite 1166/1166 (107/107 files).
